requirements.txt
Flask
mysqlclient
mysql-connector-python
Flask-WTF
Flask-Login
python-dotenv
gunicorn
